Move remove() method from api.py to remove.py
